requirements.txt
click==8.1.8
colorama==0.4.6
exceptiongroup==1.3.0
importlib-metadata==6.7.0
iniconfig==2.0.0
packaging==24.0
pika==1.3.2
pluggy==1.2.0
pytest==7.4.4
pytest-mock==3.11.1
python-dotenv==1.0.0
tomli==2.0.1
typing-extensions==4.7.1
zipp==3.15.0
